refactor: boas práticas em várias controllers

+++ api/src/main/java/vougth/api/service/TicketEventService.java
@@ -20,27 +20,18 @@ public class TicketEventService {
     @Autowired private TicketRepository ticketRepository;
     @Autowired private EventRepository eventRepository;
 
-    public ResponseEntity<Ticket> createTicket(Ticket newTicket) throws EventNotExistsException {
+    public Ticket createTicket(Ticket newTicket) throws EventNotExistsException {
         List<Event> eventList = eventRepository.findAll();
-        if (!eventList.isEmpty()){
-            ticketRepository.save(newTicket);
-            return ResponseEntity.status(201).body(newTicket);
-        }
-        return ResponseEntity.status(500).build();
+        if (!eventList.isEmpty()) ticketRepository.save(newTicket);
+        return newTicket;
     }
 
-    public ResponseEntity<List<Ticket>> getTickets(Integer id) throws TicketNotFoundException {
-        List<Ticket> ticketList = ticketRepository.findByEventCode(id);
-        return (!ticketList.isEmpty())
-                ? ResponseEntity.status(200).body(ticketList)
-                : ResponseEntity.status(204).build();
+    public List<Ticket> getTickets(Integer id) throws TicketNotFoundException {
+        return ticketRepository.findByEventCode(id);
+
     }
 
-    public ResponseEntity<Void> deleteTicketById(int id) throws TicketNotFoundException {
-        if (ticketRepository.existsById(id)) {
-            ticketRepository.deleteById(id);
-            return ResponseEntity.status(200).build();
-        }
-        return ResponseEntity.status(404).build();
+    public void deleteTicketById(int id) throws TicketNotFoundException {
+        if (ticketRepository.existsById(id)) ticketRepository.deleteById(id);
     }
 }
